requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosmtplib==4.0.1
asyncpg==0.30.0
attrs==25.3.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
frozenlist==1.6.0
google-auth==2.39.0
google-auth-oauthlib==1.2.2
gspread==6.2.0
gspread-dataframe==4.0.0
httplib2==0.22.0
idna==3.10
lxml==5.4.0
multidict==6.4.3
numpy==2.2.5
oauth2client==4.1.3
oauthlib==3.2.2
pandas==2.2.3
propcache==0.3.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycares==4.6.1
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
soupsieve==2.7
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
yarl==1.20.0
